BaseDir is now part of YamlConfig
